ECA-1716: Rudimentary support for Glassfish 2.1.1.
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@10041 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/org/ejbca/core/ejb/ca/publisher/PublisherQueueSessionBean.java
@@ -195,7 +195,7 @@ public class PublisherQueueSessionBean implements PublisherQueueSessionRemote, P
     	if(lowerBounds.length != upperBounds.length) {
     		throw new IllegalArgumentException("lowerBounds and upperBounds must have equal length");
     	}
-    	List<BigInteger> entryCountList = PublisherQueueData.findCountOfPendingEntriesForPublisher(entityManager, publisherId, lowerBounds, upperBounds);
+    	List<Integer> entryCountList = PublisherQueueData.findCountOfPendingEntriesForPublisher(entityManager, publisherId, lowerBounds, upperBounds);
     	int[] result = new int[lowerBounds.length];
 		for(int i = 0; i < lowerBounds.length && i < result.length; i++) {
 			result[i] = entryCountList.get(i).intValue();
